// workspaces/liferay-one-workspace/client-extensions/liferay-one-custom-element/src/types/salesforceProject.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export type SalesforceProject = {
	externalReferenceCode: string;
	id: number;
	koroneikiAccountKey: string;
	name: string;
	r_salesforceProjectToAccounts_accountEntryERC: string;
	r_salesforceProjectToAccounts_accountEntryId: number;
	salesforceProjectToAccountsERC: string;
};